correct warnings, separate handler templates
