+++ src/Mod/Sketcher/Gui/DrawSketchHandler.cpp
@@ -54,6 +54,7 @@
 #include "ViewProviderSketch.h"
 #include "CommandConstraints.h"
 
+#include "DrawSketchHandlerDefaultWidget.h"
 
 using namespace SketcherGui;
 using namespace Sketcher;
